move DeploymentParameters type to more relevant file
