+++ packages/core/src/new-api/types/module.ts
@@ -362,15 +362,6 @@ export interface ModuleParameters {
 }
 
 /**
- * An object containing a map of module ID's to their respective ModuleParameters.
- *
- * @beta
- */
-export interface DeploymentParameters {
-  [moduleId: string]: ModuleParameters;
-}
-
-/**
  * The results of deploying a module must be a dictionary of contract futures
  *
  * @beta
